RF NewItemTest https://github.com/orgs/RedRoverSchool/projects/3/views/1?pane=issue&itemId=90743762
