+++ src/test/java/school/redrover/NewItemTest.java
@@ -8,8 +8,6 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import school.redrover.page.freestyle.FreestyleConfigPage;
-import school.redrover.page.home.CreateNewItemPage;
 import school.redrover.page.home.HomePage;
 import school.redrover.runner.BaseTest;
 import school.redrover.runner.TestUtils;
@@ -25,26 +23,6 @@ public class NewItemTest extends BaseTest {
     private final static By NEW_ITEM_BUTTON = By.xpath("//a[@href='/view/all/newJob']");
     private final static String MESSAGE = "» This field cannot be empty, please enter a valid name";
     private final static String NEW_ITEM_NAME = "New Project";
-    private final static String FIRST_ITEM_NAME = "My_First_Project";
-    private final static String SECOND_ITEM_NAME = "My_Second_Project";
-
-    private CreateNewItemPage goToNewItemPageAndEnterName(String projectName) {
-        return new HomePage(getDriver())
-                .clickNewItem()
-                .enterItemName(projectName);
-    }
-
-    private HomePage createPipelineProject(String projectName) {
-        return goToNewItemPageAndEnterName(projectName)
-                .selectPipelineAndClickOk()
-                .gotoHomePage();
-    }
-
-    private HomePage createFreestyleProject(String projectName) {
-        return goToNewItemPageAndEnterName(projectName)
-                .selectFreestyleProjectAndClickOk()
-                .gotoHomePage();
-    }
 
     private List<String> getTextList(List<WebElement> listOfElements) {
 
@@ -55,32 +33,6 @@ public class NewItemTest extends BaseTest {
         }
 
         return textList;
-    }
-
-    @Test
-    public void testCreatePipelineFromExistingOne () {
-        List<String> itemNameList = createPipelineProject(FIRST_ITEM_NAME)
-                .clickNewItem()
-                .enterItemName(SECOND_ITEM_NAME)
-                .enterName(FIRST_ITEM_NAME)
-                .clickOkAndGoToPipelineConfigPage()
-                .gotoHomePage()
-                .getItemList();
-
-        Assert.assertTrue(itemNameList.contains(SECOND_ITEM_NAME));
-    }
-
-    @Test
-    public void testCreateFreestyleProjectFromExistingOne () {
-        List<String> itemNameList = createFreestyleProject(FIRST_ITEM_NAME)
-                .clickNewItem()
-                .enterItemName(SECOND_ITEM_NAME)
-                .enterName(FIRST_ITEM_NAME)
-                .clickOkLeadingToCertainPage(new FreestyleConfigPage(getDriver()))
-                .gotoHomePage()
-                .getItemList();
-
-        Assert.assertTrue(itemNameList.contains(SECOND_ITEM_NAME));
     }
 
     @Test
